+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -1744,6 +1744,46 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemStatusMappingObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemStatusMappingObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& statusWord,
+    const CosemByteBuffer& mappings,
+    AttributeAccessMode mutableAccess);
+  CosemStatusMappingObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& statusWord,
+    const CosemByteBuffer& mappings,
+    AttributeAccessMode mutableAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& StatusWord() const;
+  const CosemByteBuffer& Mappings() const;
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer statusWord_;
+  CosemByteBuffer mappings_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,